+++ airo-camera-toolkit/airo_camera_toolkit/utils/annotation_tool.md
