+++ airo-camera-toolkit/airo_camera_toolkit/utils/image_converter.py
